Avoids install overhead
Removes useless reload_modules: the formula itself does not need to reload salt
modules. It takes resources and time when install but it is not required.

Removes inner require: Inner require is useless since inner states are executed
in order.

Removes useless require: Docker package is not required to install docker-py.